849A/test_cdf_849A.py
import unittest
from unittest.mock import patch

from cdf_849A import CodeforcesTask849ASolution


class TestCDF849A(unittest.TestCase):
    def test_849A_acceptance_1(self):
        mock_input = ['3', '1 3 5']
        expected = 'Yes'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask849ASolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)

    def test_849A_acceptance_2(self):
        mock_input = ['5', '1 0 1 5 1']
        expected = 'Yes'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask849ASolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)

    def test_849A_acceptance_3(self):
        mock_input = ['3', '4 3 1']
        expected = 'No'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask849ASolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)

    def test_849A_acceptance_4(self):
        mock_input = ['4', '3 9 9 3']
        expected = 'No'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask849ASolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)


if __name__ == "__main__":
    unittest.main()
